Create separate requirements file for streamlit
Including `-e .` is necessary if the app is going to run on streamlit, but this then breaks the automated quartodoc publish (and I would presume pdoc too). So separate requirements file in the app subdirectory should be picked up by streamlit community cloud first as docs state it looks starting from the folder the entrypoint file is in.